Use custom SeriesOffset index type
There was a problem with the overloaded subscript operators. In the case
that I was an Int there would be 2 subscripts defined that take an int
and the compiler would have an ambiguous method error. This was a result
of splitting the concepts of Series index (which could be any
comparable) and Index used by collection. This was necessary b/c we need
to be able to define endIndex which is one passed the last item (which
doesn't make sense for a series index of type string or date).