Readd addition to create union of 2 series
When we add 2 series the result will have the labels from both series
(without duplicates). The labels that existed in both addends will have
their values added. Labels that existed in one series and not the other
will have nil for a value (but the label will still be present).